pyproject.toml
[tool.poetry]
name = "python-telegram-bot"
version = "0.1.0"
description = ""
authors = ["Roman Isakov <qe.roman@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
python-dotenv = "^0.20.0"
aiogram = "^2.21"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
